Refactoring Resources models
