--- src/FairPlaySocialSln/FairPlaySocial.Models/Localization/CultureModel.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace FairPlaySocial.Models.Localization
-{
-    /// <summary>
-    /// Holds the data for a supported culture
-    /// </summary>
-    public class CultureModel
-    {
-        /// <summary>
-        /// Short name for the culture
-        /// </summary>
-        public string? Name { get; set; }
-        /// <summary>
-        /// Display name for the culture
-        /// </summary>
-        public string? DisplayName { get; set; }
-    }
-}
