+++ src/FairPlaySocialSln/FairPlaySocial/Server/Controllers/ResourceController.cs
@@ -1,0 +1,20 @@
+﻿using FairPlaySocial.Common.CustomAttributes;
+using FairPlaySocial.Common.Global;
+using FairPlaySocial.DataAccess.Models;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+
+namespace FairPlaySocial.Server.Controllers
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    [Route("api/[controller]")]
+    [ApiController]
+    [Authorize(Roles = Constants.Roles.Admin)]
+    [ControllerOfEntity(entityName: nameof(Resource), primaryKeyType: typeof(long))]
+    public partial class ResourceController : ControllerBase
+    {
+    }
+}
