+++ src/FairPlaySocialSln/FairPlaySocial/Server/PlatformServices.cs
@@ -42,6 +42,7 @@ namespace FairPlaySocial.Server
             services.AddTransient<ProfileVisitorService>();
             services.AddTransient<ExternalReportService>();
             services.AddTransient<CultureService>();
+            services.AddTransient<ResourceService>();
             services.ConfigureIpDataService(configuration);
             services.ConfigureIpStackService(configuration);
         }
